🛠️  Refactor: 로직 원심분리기 작동
